chore(npm): update package-lock format (#58)
update the format of package-lock.json to use v2, which was introducted
in npm 7. the file was updated by running `npm install` using npm
v8.0.0, as there is no guidance on
https://docs.npmjs.com/cli/v7/configuring-npm/package-lock-json for
doing a proper update to v2 of the lock file.

this comes at the risk of updating some of our dependencies underneath
us, but since the library has not yet reached v1.0, the author of this
pr is considering that to be an acceptable risk (and may in fact help us
galvanize our set up).