fix(react): squash react version not specified warning
this commit squashes the following warning message that is printed to
the console today:

> Warning: React version not specified in eslint-plugin-react settings.
> See https://github.com/yannickcr/eslint-plugin-react#configuration

because @stencil/eslint uses eslint-plugin-react for JSX related
linting. this commit puts an invalid semver string in to value field
for the react version. admittedly it's a bit of a hack, but it does
manage to squash the error message.