chore(repo): move repo to community (#82)
this commit performs a variety of updates on the repo in order to
prepare it for being moved to the community.

the project has been renamed from `@stencil/eslint-plugin` to
`@stencil-community/eslint-plugin`. this change in scope allows us to be
clear that this package is no longer maintained by the stencil core
team.

various urls have been updated to point to the stencil-community repo
version of this project. in the event where it made sense to continue to
point back to a stencil/ionic owned page (e.g. a specific code of
conduct does not exist in the community), it has been left as is.

all github files that were specific to the ionic-team github
organization were removed