fix(lint): update scopes for all rules (#83)
this commit fixes the scope of each rule. previously, each rule was
under the `@stencil` scope when this project was being maintained by the
stencil core team. the project has since begun to move to the community
model, where the npm package is published under the
`@stencil-community/` scope. in the original migration to the community
model (fb8dab7), updating the scopes of each rule was missed. this
caused lint rule lookups to fail.

by updating the scopes for each rule, the predefined configs, and the
docs, rules can once again be used